Move extension.d.ts back
